+++ helpers/controllersWrapper.js
@@ -1,0 +1,13 @@
+const controllersWrapper = (ctrl) => {
+  const func = async (req, res, next) => {
+    try {
+      await ctrl(req, res, next);
+    } catch (error) {
+      next(error);
+    }
+  };
+
+  return func;
+};
+
+module.exports = controllersWrapper;
