+++ models/contacts.json
@@ -46,11 +46,5 @@
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
-  },
-  {
-    "id": "Ed4dJ_hhIajAFkYLkQ19",
-    "name": "test",
-    "email": "test@mail.com",
-    "phone": "1111111111111"
   }
 ]
