+++ routes/api/contacts.js
@@ -1,95 +1,22 @@
 const express = require("express");
 const {
-  getContactList,
-  getContactById,
-  addContact,
-  updateContact,
-  removeContact,
-} = require("../../models/contacts");
-const Joi = require("joi");
-const { HttpError } = require("../../helpers");
+  getAll,
+  getById,
+  add,
+  update,
+  remove,
+} = require("../../controllers/contacts");
 
 const router = express.Router();
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
+router.get("/", getAll);
 
-router.get("/", async (_, res, next) => {
-  try {
-    const result = await getContactList();
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:id", getById);
 
-router.get("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await getContactById(id);
+router.post("/", add);
 
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
+router.put("/:id", update);
 
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-
-    if (error) {
-      throw HttpError(400, "Missing required name field");
-    }
-
-    const result = await addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put("/:id", async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-
-    if (error) {
-      throw HttpError(400, "Missing fields");
-    }
-
-    const { id } = req.params;
-    const result = await updateContact(id, req.body);
-
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.delete("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await removeContact(id);
-
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-
-    res.json({ message: "Contact deleted" });
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:id", remove);
 
 module.exports = router;
